Define a domain instead of create & launch it - support for modifying disk images
QEMU doesn't seem to notice if you externally modify the disk image of a running virtual machine.  I was hoping that if a VM is started paused that I might be able to inject files into it, but they ultimately never showed up when the VM booted.

Maybe it's a caching problem?